ANDROID: add script for reporting on CKI coverage
The common kernel interface (CKI) is the set of kernel interfaces required
for the Android Open Source Project to run. Many of these interfaces are
system calls.

Add a script which reports the set of system calls in the CKI and also
the current LTP test coverage on them in VTS.

Bug: 37480412
Change-Id: I305ce3f530ee299a33416977a7bf59ddfdeaba6c
Signed-off-by: Steve Muckle <smuckle@google.com>