android: fix make_perser to accept 'ar' arguments without hyphens
The ar command line parser broke due to new way of running ranlib on
archives - 'ar rcs libfoo.a foo1.o foo2.o ..'. The fix assumes if arg
parser failed to find any arguments (with hyphens) it is this new way
of creating archieves and simply looks for the file with ".a" extension.

NOTE: This prepares for the upcoming LTP release merge and so the
parsers and build generators may or may not work for older makefile
output

Change-Id: I8774249002100a26041f8072f17cd84cd8e1d73a
Signed-off-by: Sandeep Patil <sspatil@google.com>