android: fix binaries built out of .c and .o in single gcc invocation
The new LTP update adds several tests that are built using a mix of .c
and .o files listed on a single gcc commandline. The build generator by
default considers these "source" files and generators the .bp fragments
with .o files listed in sources which fails the build.

Fix the generator so if it sees the .o files in the cc_compilelink
sources, it then goes and finds the "real" sources for those .o
files from the cc_compile map

Test: 'ltp_iogen' module is built correctly with the fixed .bp fragments

Change-Id: Ica9980d3f2c112b4ee594865d3b8c016d37c18a9
Signed-off-by: Sandeep Patil <sspatil@google.com>