android: pty01: Fix pty01 test for Android.
The test fails because 'grantpt()' is a no-op in bionic and doesn't set
the /dev/pts/X mode to '020600' as expected by the test. The change
skips that check if __BIONIC__ is defined so the rest of the test(s)
can proceed and detect other failures if any.

Bug: 65053723
Test: run vts-kernel -m VtsKernelLtp -t pty.pty01

Signed-off-by: Sandeep Patil <sspatil@google.com>
Acked-by: Petr Vorel <pvorel@suse.cz>
(cherry picked from commit 53eed736f1655217ff7ebc1757383ef1d134eab4)

Change-Id: Ied7498dca9232b4ff921ca387eaf781e0909806b