Enable all tests that were disabled due to disabled cflags
There were 62 tests disabled due to the -DTST_USE_COMPAT16_SYSCALL=1
CFLAG for thier builds. Paul Lawrence and I confirmed that the
compat 16 tests (with "_16") extension should pass, so I'm
re-enabling all those tests.

I ran the 32-bit versions of all of them and 14 of them fail. I filed
bugs against those 14.

I also discovered that the "_16" versions of the tests are not
being built since the generation scripts are not adding them
to android/ltp_package_list.mk. I filed a bug for this as well
(b/73166481).

Bug: 31226478 (and many more)
Test: ran all the tests enabled by this CL
Change-Id: I34b28709e30c7a07621ed067d454e3818e56a6d6
Signed-off-by: Todd Kjos <tkjos@google.com>