UPSTREAM: {rt_, }sigsuspend01: Fix build for non-glibc platforms
The tests dig inside the sigset_t to check if the kernel reverted any
changes made to the signal mask. This is done by comparing the first
32 signal flag bits by dereferencing the sigset_t and reading __val[0],
which is an unsigned long. If the field is not called __val (with
Android's bionic it is called '__bits' instead) the build breaks.

Change the tests to use memcmp for the first sizeof(unsigned long) bits,
which should be equivalent. This makes the test more portable.

Signed-off-by: Alistair Strachan <astrachan@google.com>
Signed-off-by: Petr Vorel <pvorel@suse.cz>
(cherry picked from commit ddee74b92fca5b1a835d363b0ddcb2e848ef7b83)
Bug: 37950676
Change-Id: I8b63bf497029245b757bdac6f95cc1eb5489dba8
Signed-off-by: Alistair Strachan <astrachan@google.com>