ANDROID: cki_coverage: check all arch syscalls, remove result parsing
The CKI now includes all syscalls supported by the Linux kernel. Instead
of looking at bionic files to determine the CKI, download the required
Linux source files and determine the CKI set from that.

The default is now to provide the combined CKI across all architectures.
It is still possible to restrict output to a single architecture with
-a/--arch.

Remove the test result parsing as it has not been used and introduces
unnecessary complexity.

Bug: 79883734
Change-Id: I06d889c3306fcad1d6d2a08bb9b7a5b7154b6b1f