ANDROID: cki_coverage: tweak fadvise64 special case
bionic implements posix_fadvise() using one of three fadvise64 variants,
depending on the target.  The way these syscalls are exported to
userspace is arch-dependent and confusing:

- 32-bit x86 exports two variants, and bionic uses the latter
  exclusively:

  fadvise64(int fd, s64 offset, s32 len, int advise)
  fadvise64_64(int fd, s64 offset, s64 len, int advise)

- ARM64 and x86_64 each export one variant that's named fadvise64 in the
  UAPI headers, but use a 64-bit `len' like x86 fadvise64_64.

- 32-bit ARM exports a variant of fadvise64_64 with reordered arguments:

  arm_fadvise64_64(int fd, int advice, s64 offset, s64 len)

IMO there's no good way to represent this in a coverage table, but the
least terrible option is to say fadvise64 is uncovered.  Of all these
variants, 32-bit x86 fadvise64 is the only one that isn't covered by
LTP.

Bug: 112469908
Test: run cki_coverage.py, inspect output
Change-Id: I1994f9f55398a44a944da3fc98ada0d4d64a39de
Signed-off-by: Greg Hackmann <ghackmann@google.com>