getrlimit/getrlimit03: fix breakage with Aarch32
A 64-bit ARM kernel running an Aarch32 user binary has the ugetrlimit
and prlimit64 syscalls, but not the getrlimit syscall. This test
currently assumes getrlimit is always available.

The same happens for ppc64le and quite likely for any architecture Linux
was ported on recently.

Modify the test so that it does not fail in this configuration.

Signed-off-by: Steve Muckle <smuckle@google.com>
Signed-off-by: Richard Palethorpe <rpalethorpe@suse.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
(cherry picked from commit fac783b5d6631efa70d2ea3257142763d944ba32)
Bug: 112483925
Test: run vts-kernel -m VtsKernelLtp -t syscalls.getrlimit03
Signed-off-by: Steve Muckle <smuckle@google.com>
Change-Id: I64f52e3a09ec8e01e974661ad4974d0c18cbaeaa