ANDROID: cki_coverage: fix reference to VTS LTP stable list
The VTS LTP stable test list was recently changed from a list of test
names to a list of tuples, where each tuple contains the test name and a
boolean indicating whether the test is mandatory.

Fix cki_coverage.py to continue to refer to the test names in this list.

Test: verify cki_coverage.py output
Change-Id: Iff7abd8cf755e35e5af04b1e312d0529bd40703b
Signed-off-by: Steve Muckle <smuckle@google.com>