android_build_generator: sort autogenerated dependencies
Whenever the test configuration is changed, it is not clear how the
generated files are going to be updated with added/deleted entries.
This patch orders the header and source files that are imported, as well
as the flags, so that the generated files are deterministic,
simplifying the tracking of their versioning.

Resulting gen.bp updated with reordered headers, sources and flags.

Test: manual generated files inspection, build
Change-Id: I39bba83143616393e25b54296964259562b73530
Signed-off-by: Alessio Balsini <balsini@google.com>