UPSTREAM: aio_tio: determine alignment based on target block device
The alignment for O_DIRECT operations has to match the blocksize of the
underlying block device. Determine the alignment from the target file.

aio_tio test cases 1 and 2 failed (nondeterministic) on aarch64 when run
on a 4096 byte blocksize and with an alignment of 512 bytes.

Determining the blocksize from the block device and using it for the
alignment resolves the test issue.

Bug: 121039294
Reviewed-by: Alessio Balsini <balsini@google.com>
Reviewed-by: Steve Muckle <smuckle@google.com>
Signed-off-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: I66cee31660f059289d7e0ffa8674179bd49059fe
(cherry picked from commit a40e1c065ababeec702622cedc6a8987d00b01ae)