setpriority01: Skip only PRIO_USER when unable to add test user
We don't need to skip all the tests just because we are unable to add
a test user. Not having a test user only affects PRIO_USER test case.
So just skip that one and continue running the rest of the tests when
useradd is not available or can't write to /etc/passwd (Eg: read-only
filesystem)

If useradd is present and it still fails for other reasons, then
consider it an error in the test preparation phase.

This also allows this test case to be built and run on Android.

Signed-off-by: Saravana Kannan <saravanak@google.com>
Acked-by: Cyril Hrubis <chrubis@suse.cz>
(cherry picked from commit db82b596ccfb08470f2f5bf19b3d9cee7b2e1a9e)
Bug: 112484634
Change-Id: Id9c993125a22792a161c0b7e639694486492f390
Test: run vts-kernel -m VtsKernelLtp
Signed-off-by: Steve Muckle <smuckle@google.com>