ANDROID: remove build rules for old select variant tests
An old out of tree patch had added support to test variants of the
select syscall (2a920f788424, "select: tests offer choice of syscall
path"). That has been superseded by an upstream version so remove the
build rules for the old targets (the rest of the old patch is already
gone).

Change-Id: I1188363cf71ae45d6cdc7066e8f75ed7fa93b26f
Signed-off-by: Steve Muckle <smuckle@google.com>