eas: sched_boost: give time for utilization to track task usage
PELT will set the initial util_avg of a task to half the spare capacity
of the CPU the task starts to run on. This will cause problems for this
test since a task which is expected to be small will potenitally show a
large utilization while we are monitoring it.

Take half the test time (currently one second) to allow util_avg to
track the actual task usage prior to starting the trace and tracking the
max util_avg for comparison.

Bug: 144809570
Change-Id: Ie5d519078707749d94011dde3f13df7810955c7b
Signed-off-by: Steve Muckle <smuckle@google.com>